Support sqlite as a connection type (#11)
* Recreated the migrations to store the url of the connection instead of the individual components

* Use tui-menu for database type and environment dropdowns in new connection modal

* Added logic to query the sqlite db

* Added nightly rust compiler and just file to run unstable rustfmt rules

* Clippy lints

* Fixed column width when there are no values for it

* Focus sqlite navigation and display to remove redundant navigation and views

* Show app information when no connection has been selected

* Only delete when enter key is pressed

* Fixed modal action button navigation

* Enhanced new connection flow

+++ crates/d7s_ui/src/handlers/key_event.rs
@@ -101,7 +101,7 @@ pub fn handle_connection_list_navigation<T: TableData + Clone>(
                 table_state.view.state.select_column(Some(last_col));
                 table_state.adjust_offset_for_selected_column(
                     last_col,
-                    DEFAULT_TERMINAL_WIDTH,
+                    DEFAULT_TERMINAL_WIDTH as usize,
                 );
                 TableNavigationHandler::wrap_columns(table_state);
             }
